Fix #4413: Update placeholder launcher icon (#4576)
## Explanation
Fixes #4413

This PR updates the placeholder launcher icon with something that better fits the adaptive icon expectations.

Note that this PR is removing the rounded icon per the advice of the adaptive icon support page (that it should be preferred not to be used unless necessary), and it's removing the play store icon since it shouldn't be needed. It's also inverting the colors of the icon since that seems to be less visually jarring for the circle-in-circle icon cases.

Note that this PR is introducing more binary files to the history, but they're small. Longer term we can look into rendering these PNGs during build time, but it's not an option right now.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If this PR fixes part of an issue, prefix the title with "Fix part of #bugnum: ...".)
- [x] Any changes to [scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets) files have their rationale included in the PR explanation.
- [x] The PR follows the [style guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android Studio ([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers ([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
This section will be filled in post-merge (#4567 is tracking ensuring that the description is finished later).