Rework e2e tests, and make them work in Bazel.
The main changes for compatibility are using Python 2 and using a Nexus
5X device instead of generic phone for the emulator. I also added some
additional nice-to-haves in .bazelrc.

From the testing side, I realized that the tests were sensitive to
display density so I wanted to start adding scrolling stability (like we
do in Espresso). Unfortunately, UiAutomator makes this difficult with
two different, and mostly incompatible APIs. I abstracted these into a
new layer, and added a DSL on top of it to make test writing really
clean. I also fixed an accidental slow-down due to waiting
configurations. That being said, I've noticed that the end of the test
is really flaky on the 'Return to Topic' button (and sort of was
before), and I have no idea nor visibility into why.

I haven't verified the image region test yet, and there's additional
cleanup work needed here.