Set custom protobuf artifact for m1 architecture
Changed the conditional check for the 'protobuf_platform' property from checking for property presence to checking for the current os. This is because `project.rootProject.hasProperty('protobuf_platform')` does not actually read the properties file as it is meant to, therefore the if condition was never met and the artifact was never set correctly.

Changed how the property is read by loading the local.properties file with inputstream.