Fix #4684: Revert (most of) #4587 (#4686)
## Explanation
Fixes #4684

This PR reverts #4587. I think the change of match_parent to
wraps_content caused the promoted story container to grow for larger
images, and this wasn't caught until we tried the changes with
production assets. I noticed other places in the PR that might have the
same issue, so rather than trying to make sense of everything in order
to properly fix it & add tests, I opted to revert the change since the
regression is beta-blocking (even though it wasn't a clean reversion and
required manual changes).

Note that the reversion includes keeping the changes to
recently_played_fragment since otherwise changes in
RecentlyPlayedFragmentTest fail, and the change seems unlikely to cause
any problems with respect to potential thumbnail sizing issues.

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
There's not much to show here--see the issue for context on the problem.
The post-reversion just reverts things back to their previous state
(i.e. no regression).