Fix #4746: Update app to target SDK 31 (#4747)
## Explanation
Fixes #4746

This PR updates the app to target SDK 31 instead of 30 as this is now
required for continued Play Store releases.

Verification from the PR standpoint heavily relies on CI checks passing.
Beyond that, manual testing is needed to ensure no Android behavior
changes have led to regressions or new breakages.

This PR does NOT include updating Robolectric tests to run on SDK 31
since it's actually much harder than I had expected, and would lead to
potential production-affecting changes (as it would require updating
Kotlin & potentially other dependencies). #4748 is tracking fixing this
in the long-term.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A, mostly. While this is an infrastructure change, Android component
behaviors change with different target SDKs so it's quite possible for
there to be UI regressions or issues. Fortunately, most of the UI used
by the app is material and Jetpack (meaning it ships with the app and
will be mostly unaffected by target SDK changes).