Add support for Bazel Espresso tests.
This uses a new, custom instrumentation runner that will later be used
in e2e tests. This provides a few advantages over Bazel's:
- Interop with Oppia's custom managed AVDs (which means faster
start-up).
- Interop with Oppia's emulator runner (which means the ability to
customize rendering and acceleration settings).

The runner is designed to attach to a matching emulator currently
running, or (eventually) to a specific serial when running with a real
device (this isn't fully implemented yet, but will be easy to add in the
future). If a matching emulator isn't already running, an ephemeral one
will be created (using an existing AVD, so it's a bit faster than the
Bazel alternative; maybe about 2x faster). No effort is done to manage
multiple tests trying to use the same emulator (it's expected that this
will eventually be handled at a higher level during test suite
orchestration).

The only Espresso test enabled is SplashActivityTest, and it's done
haphazardly. Some visibilities needed to be exposed within
app/BUILD.bazel, and Mockito needed to be removed from some test
dependencies due to an incompatibility with the legacy dexer used by
Bazel. These are temporary workarounds until the app is moved to a newer
version of Bazel that avoids needing to depend on legacy versions.

It may be prudent to wait until the Bazel migration finishes to enable
Espresso tests broadly for a few reasons:
- Exposed visibilities (per previous paragraph) are a bit messy.
- Depending on //app:app results in *all* production activity manifest
declarations being retained (this commit introduces a custom generator
to try and manage this, but it's mostly a stopgap).
- Espresso tests are *extremely* slow (~35x slower than Robolectric
equivalents), and likely don't need to be run at the same rate as
Robolectric. A rethink of overall app test strategy is needed to make a
determination on what situations actually necessitate Espresso tests,
and how to declare these to keep test runtimes lower.

Finally, this is a good place to mention that instrumentation mode can't
be used for test binaries since they rely on R resources. I think it may
be possible to work around this by inlining R resources within a test,
but it's tricky and needs more investigation. This won't be an issue for
e2e tests since they'll never have hidden information about the app.